move hosting into a separate file
